Accach comments to correct case clauses...
When attaching comment, before searching backwards over newlines, search forwards to see if we meet
a case / comm clause start decoration point. If we find this, compare it's indent with the indent of
the comment. If it's the same, attach straight away. If it's not the same, attach backwards. Prevent
any more forward searching steps because we never want to attach forward.